+++ src/ios/FirebaseCrashPlugin.m
@@ -53,4 +53,15 @@
     [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
 }
 
+
+- (void)forceCrash:(CDVInvokedUrlCommand *)command {
+    
+    @[][1];
+    
+    CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
+    [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+}
+
+
+
 @end